+++ Core/Networking/Http/HttpClientHeaderModifier.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -15,17 +16,20 @@ namespace Unity.Cloud.Common
     {
         readonly IHttpClient m_BaseClient;
         readonly Dictionary<string, string> m_Headers;
+        readonly string m_UrlFilter;
 
         /// <summary>
         /// Creates and instance of <see cref="HttpClientHeaderModifier"/>.
         /// </summary>
         /// <param name="httpClient">The client who's requests will have headers added.</param>
         /// <param name="headers">The headers to add to each request.</param>
+        /// <param name="urlFilter">The optional url filter to determine which requests should have the headers added. A null or empty filter will add the headers to all requests.</param>
         /// <exception cref="ArgumentNullException">Thrown if <paramref name="httpClient"/> or any key in <paramref name="headers"/> is null.</exception>
-        public HttpClientHeaderModifier(IHttpClient httpClient, Dictionary<string, string> headers)
+        public HttpClientHeaderModifier(IHttpClient httpClient, Dictionary<string, string> headers, string urlFilter = null)
         {
             m_BaseClient = httpClient ?? throw new ArgumentNullException(nameof(httpClient));
             m_Headers = headers != null && !headers.Keys.Any(string.IsNullOrWhiteSpace) ? headers : throw new ArgumentNullException(nameof(headers), $"A key in {nameof(headers)} is null or white space.");
+            m_UrlFilter = urlFilter;
         }
 
         /// <inheritdoc />
@@ -45,9 +49,12 @@ namespace Unity.Cloud.Common
 
         protected void AddHeaders(HttpRequestMessage request)
         {
-            foreach (var header in m_Headers)
+            if (string.IsNullOrEmpty(m_UrlFilter)|| Regex.IsMatch(request.RequestUri.ToString(), m_UrlFilter, RegexOptions.IgnoreCase))
             {
-                request.Headers.Add(header.Key, header.Value);
+                foreach (var header in m_Headers)
+                {
+                    request.Headers.Add(header.Key, header.Value);
+                }
             }
         }
     }